fix: repair android package and window startup
