+++ src-tauri/gen/android/app/src/main/java/moe/sable/client/MainActivity.kt
@@ -1,4 +1,4 @@
-package moe.sable.app
+package moe.sable.client
 
 import android.graphics.Bitmap
 import android.os.Bundle
