+++ src-tauri/gen/android/buildSrc/src/main/java/moe/sable/client/kotlin/RustPlugin.kt
